fix: Centralize most of the webpack configuration.
Fixes webpack errors introduced in 0547bd5dd8093819d2cf5f9346648e1a2773801f.
