+++ .eslintrc.yaml
@@ -53,6 +53,9 @@ globals:
   __POSTS_APP_URL__: false
   __PHOTOS_APP_URL__: false
   __RESUME_APP_URL__: false
+  __PHOTOS_SERVICE_URL__: false
+  __WORDS_SERVICE_URL__: false
+  __POSTS_SERVICE_URL__: false
   __CLIENT_URL__: false
   __PUBLISHED_RESUME_URL__: false
   __CAMPAIGN_SOURCE__: false
