+++ babel.register.js
@@ -1,8 +1,9 @@
 const path = require("path");
+const util = require("./util");
 
 require("@babel/register")({
     configFile: path.join(__dirname, "babel.config.js"),
     only: [
-        /(?:node_modules\/(@randy\.tarampi|query-string|strict-uri-encode|strip-ansi|ansi-regex)|me\/packages)\//
+        util.babelRegisterInclusions
     ]
 });
