+++ packages/jsx/gulpfile.js
@@ -48,7 +48,7 @@ gulp.task("clean", (callback) => {
 
 gulp.task("webpack", function (callback) {
     const WebPack = require("webpack");
-    const WebPackConfig = require("./webpack.config");
+    const WebPackConfig = require("./webpack.client.config");
     const webpackConfig = Object.assign({}, WebPackConfig, {
         mode: "production"
     });
@@ -61,7 +61,7 @@ gulp.task("webpack", function (callback) {
 
 gulp.task("webpack:dev", (callback) => {
     const WebPack = require("webpack");
-    const WebPackConfig = require("./webpack.config");
+    const WebPackConfig = require("./webpack.client.config");
     const webpackConfig = Object.assign({}, WebPackConfig);
 
     WebPack(webpackConfig, function (err, stats) {
