+++ packages/letter/package.json
@@ -24,7 +24,7 @@
     "postinstall": "npm run build",
     "start": "npm run dev",
     "dev": "npm run dev:client",
-    "dev:client": "rm -rf dist/*.js && npx webpack-serve --hotClient",
+    "dev:client": "rm -rf dist/*.js && npx webpack-serve ./webpack.client.config.js --hotClient",
     "build": "npx gulp -LLLL --color build && npm run letter",
     "start:express": "npm run kill:express; node ./express.js & echo $! > /tmp/ca.randytarampi.me.letter.express.pid; echo \"Started express (`cat /tmp/ca.randytarampi.me.letter.express.pid`) from /tmp/ca.randytarampi.me.letter.express.pid\";",
     "kill:express": "if [ -f /tmp/ca.randytarampi.me.letter.express.pid ]; then echo \"Killing express (`cat /tmp/ca.randytarampi.me.letter.express.pid`) from /tmp/ca.randytarampi.me.letter.express.pid\"; kill -9 `cat /tmp/ca.randytarampi.me.letter.express.pid`; rm /tmp/ca.randytarampi.me.letter.express.pid; fi;",
