+++ packages/letter/webpack.client.config.js
@@ -1,0 +1,10 @@
+const path = require("path");
+const webpackBaseConfig = require("../../webpack.client.config.base");
+
+module.exports = webpackBaseConfig({
+    sourceDirectoryPath: __dirname,
+    compliationDirectoryPath: path.join(__dirname, "dist"),
+    entry: {
+        letter: ["@babel/polyfill", path.join(__dirname, "./public/views/index.jsx")]
+    }
+});
