+++ packages/resume/package.json
@@ -25,7 +25,7 @@
     "postinstall": "npm run build",
     "start": "npm run dev",
     "dev": "npm run dev:client",
-    "dev:client": "rm -rf dist/*.js && npx webpack-serve --hotClient",
+    "dev:client": "rm -rf dist/*.js && npx webpack-serve ./webpack.client.config.js --hotClient",
     "build": "npx gulp -LLLL --color build && npm run resume",
     "resume": "npm run resume:json && npm run resume:pdf && npm run resume:html",
     "start:express": "npm run kill:express; node ./express.js & echo $! > /tmp/ca.randytarampi.me.resume.express.pid; echo \"Started express (`cat /tmp/ca.randytarampi.me.resume.express.pid`) from /tmp/ca.randytarampi.me.resume.express.pid\";",
