+++ packages/www/gulpfile.js
@@ -100,7 +100,7 @@ gulp.task("styles", gulp.series(["styles:dev"]), () => {
 
 gulp.task("webpack:dev", function (callback) {
     const WebPack = require("webpack");
-    const WebPackConfig = require("./webpack.config");
+    const WebPackConfig = require("./webpack.client.config");
     const webpackConfig = Object.assign({}, WebPackConfig);
 
     WebPack(webpackConfig, function (err, stats) {
@@ -111,7 +111,7 @@ gulp.task("webpack:dev", function (callback) {
 
 gulp.task("webpack", function (callback) {
     const WebPack = require("webpack");
-    const WebPackConfig = require("./webpack.config");
+    const WebPackConfig = require("./webpack.client.config");
     const webpackConfig = Object.assign({}, WebPackConfig, {
         mode: "production"
     });
