+++ util.js
@@ -1,0 +1,4 @@
+module.exports = {
+    babelLoaderExclusions: /\/node_modules\/(?!@randy\.tarampi|query-string|strict-uri-encode|strip-ansi|ansi-regex)\//,
+    babelRegisterInclusions: /\/(?:node_modules\/(?:@randy\.tarampi|query-string|strict-uri-encode|strip-ansi|ansi-regex)|packages)\//
+};
